Pipelines: fix jobs needs lists
Each jobs needs list is supposed to only contain direct dependencies
for scheduling purposes, unless "enable-artifacts-buildcache" is
specified.  Only in that case are the needs lists supposed to contain
all transitive dependencies.  This changes fixes a bug that caused the
needs lists to always contain all transitive dependencies, regardless
of whether or not "enable-artifacts-buildcache" is specified.